pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "venvoy"
version = "0.1.0"
description = "A multi-OS, multi-architecture, immutable, portable, and shareable AI-ready python environment with ultra-fast package managers"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "zaphodbeeblebrox3rd", email = "zaphodbeeblebrox3rd@users.noreply.github.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Systems Administration",
]
requires-python = ">=3.9"
dependencies = [
    "click>=8.0.0",
    "docker>=6.0.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "packaging>=21.0",
    "requests>=2.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
    "pytest-mock>=3.10.0",
    "pytest-docker>=2.0.0",
]

[project.scripts]
venvoy = "venvoy.cli:main"

[project.urls]
Homepage = "https://github.com/zaphodbeeblebrox3rd/venvoy"
Repository = "https://github.com/zaphodbeeblebrox3rd/venvoy"
Issues = "https://github.com/zaphodbeeblebrox3rd/venvoy/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 120
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 120

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.flake8]
max-line-length = 120
extend-ignore = ["E203", "W503"]  # E203: whitespace before ':', W503: line break before binary operator 